requirements.txt
blinker==1.8.2
Bootstrap-Flask==2.2.0
click==8.1.7
Flask==2.3.2
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.12.2
Werkzeug==3.0.0
WTForms==3.0.1
